+++ e2e/resolve-async/some-other-file.js
@@ -5,8 +5,4 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import type {Resolver} from '../defaultResolver';
-
-declare const userResolver: Resolver;
-
-export default userResolver;
+export default 'Hello from mapped file!!';